+++ W3ChampionsStatisticService/WebApi/ActionFilters/W3CAuthenticationService.cs
@@ -6,6 +6,7 @@ using System.Text.RegularExpressions;
 using Microsoft.IdentityModel.Tokens;
 using System.Collections.Generic;
 using System.Diagnostics;
+using W3ChampionsStatisticService.Services.Tracing;
 
 namespace W3ChampionsStatisticService.WebApi.ActionFilters;
